refactor(tuners): extract worker-metric collection into worker_metrics module
Move collect_system_metrics, fetch_pg_stat_database_snapshot, and
compute_io_metrics out of orchestrator.py into a dedicated worker_metrics
module as free functions taking explicit handles (env, connect/disconnect).
The orchestrator keeps thin delegating methods so the test-pinned and
internal call surface is unchanged.
